Move couch compact files from services to libs
